+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Microsoft 365/Extra Themes/PaletteMicrosoft365BlackDarkMode.cs
@@ -410,7 +410,7 @@ namespace Krypton.Toolkit
     /// Provides a base for Office 365 palettes.
     /// </summary>
     /// <seealso cref="PaletteBase" />
-    public abstract class PaletteMicrosoft365BlackDarkModeBase : PaletteBase 
+    public abstract class PaletteMicrosoft365BlackDarkModeBase : PaletteBase
     {
         #region Static Fields
 
@@ -532,17 +532,17 @@ namespace Krypton.Toolkit
                                                             };
 
         private static readonly Color[] _appButtonTrack = { Color.FromArgb(255, 251, 230),
-                                                            Color.FromArgb(248, 230, 143),
-                                                            Color.FromArgb(238, 213, 126),
-                                                            Color.FromArgb(254, 247, 129),
-                                                            Color.FromArgb(240, 201, 41)
+                                                            Color.FromArgb(178, 178, 178),
+                                                            Color.FromArgb(176, 176, 176),
+                                                            Color.FromArgb(179, 179, 179),
+                                                            Color.FromArgb(160, 160, 160)
                                                           };
 
         private static readonly Color[] _appButtonPressed = { Color.FromArgb(235, 227, 196),
-                                                              Color.FromArgb(228, 198, 149),
-                                                              Color.FromArgb(166, 97, 7),
-                                                              Color.FromArgb(242, 155, 57),
-                                                              Color.FromArgb(236, 136, 9)
+                                                              Color.FromArgb(185, 185, 185),
+                                                              Color.FromArgb(35, 35, 35),
+                                                              Color.FromArgb(50, 50, 50),
+                                                              Color.FromArgb(100, 100, 100)
                                                             };
 
         private static readonly Color[] _buttonBorderColors = { Color.FromArgb(180, 180, 180), // Button, Disabled, Border
@@ -557,7 +557,7 @@ namespace Krypton.Toolkit
         private static readonly Color[] _buttonBackColors = { Color.FromArgb(250, 250, 250), // Button, Disabled, Back 1
                                                               Color.FromArgb(250, 250, 250), // Button, Disabled, Back 2
                                                               Color.FromArgb(91, 91, 91), // Button, Tracking, Back 1
-                                                              Color.FromArgb(89, 89, 89), // Button, Tracking, Back 2
+                                                              Color.FromArgb(33, 33, 33), // Button, Tracking, Back 2
                                                               Color.FromArgb(121, 121, 121), // Button, Pressed, Back 1
                                                               Color.FromArgb(91, 91, 91),  // Button, Pressed, Back 2
                                                               Color.FromArgb(91, 91, 91), // Button, Checked, Back 1
@@ -565,6 +565,7 @@ namespace Krypton.Toolkit
                                                               Color.FromArgb(41, 41, 41), // Button, Checked Tracking, Back 1
                                                               Color.FromArgb(70, 70, 70)  // Button, Checked Tracking, Back 2
                                                             };
+
         #endregion
 
         #region Variables