+++ Source/Krypton Components/Krypton.Toolkit/Palette Builtin/Microsoft 365/Extra Themes/PaletteMicrosoft365SilverDarkMode.cs
@@ -25,7 +25,7 @@ namespace Krypton.Toolkit
         #endregion
 
         #region Images
-        
+
         private static readonly Image[] _radioButtonArray;
 
         private static readonly Image _silverDropDownButton = Office2010Arrows._2010BlueDropDownButton;
@@ -406,7 +406,7 @@ namespace Krypton.Toolkit
     /// Provides a base for Office 365 palettes.
     /// </summary>
     /// <seealso cref="PaletteBase" />
-    public abstract class PaletteMicrosoft365SilverDarkModeBase : PaletteBase 
+    public abstract class PaletteMicrosoft365SilverDarkModeBase : PaletteBase
     {
         #region Static Fields
 
@@ -521,27 +521,56 @@ namespace Krypton.Toolkit
 
         #region Colour Arrays
 
-        private static readonly Color[] _appButtonNormal = { Color.FromArgb(243, 245, 248), Color.FromArgb(214, 220, 231), Color.FromArgb(188, 198, 211), Color.FromArgb(254, 254, 255), Color.FromArgb(206, 213, 225) };
-        private static readonly Color[] _appButtonTrack = { Color.FromArgb(255, 251, 230), Color.FromArgb(248, 230, 143), Color.FromArgb(238, 213, 126), Color.FromArgb(254, 247, 129), Color.FromArgb(240, 201, 41) };
-        private static readonly Color[] _appButtonPressed = { Color.FromArgb(235, 227, 196), Color.FromArgb(228, 198, 149), Color.FromArgb(166, 97, 7), Color.FromArgb(242, 155, 57), Color.FromArgb(236, 136, 9) };
-        private static readonly Color[] _buttonBorderColors = { Color.FromArgb(212, 212, 212), // Button, Disabled, Border
-                                                                Color.FromArgb(221, 207, 155), // Button, Tracking, Border 1
-                                                                Color.FromArgb(192, 167, 119), // Button, Tracking, Border 2
-                                                                Color.FromArgb(147, 125,  90), // Button, Pressed, Border 1
-                                                                Color.FromArgb(255, 196,  68), // Button, Pressed, Border 2
-                                                                Color.FromArgb(158, 130,  85), // Button, Checked, Border 1
-                                                                Color.FromArgb(254, 218, 144)  // Button, Checked, Border 2
+        private static readonly Color[] _appButtonNormal =
+        {
+            Color.FromArgb(243, 245, 248),
+            Color.FromArgb(214, 220, 231),
+            Color.FromArgb(188, 198, 211),
+            Color.FromArgb(254, 254, 255),
+            Color.FromArgb(206, 213, 225)
         };
-        private static readonly Color[] _buttonBackColors = {   Color.FromArgb(221, 221, 221), // Button, Disabled, Back 1
-                                                                Color.FromArgb(236, 236, 236), // Button, Disabled, Back 2
-                                                                Color.FromArgb(54,64,88), // Button, Tracking, Back 1
-                                                                Color.FromArgb(106,123,164), // Button, Tracking, Back 2
-                                                                Color.FromArgb(95, 107, 137), // Button, Pressed, Back 1
-                                                                Color.FromArgb(54,64,88), // Button, Pressed, Back 2
-                                                                Color.FromArgb(54,64,88), // Button, Checked, Back 1
-                                                                Color.FromArgb(122, 137, 174), // Button, Checked, Back 2
-                                                                Color.FromArgb(119, 134, 172), // Button, Checked Tracking, Back 1
-                                                                Color.FromArgb(135, 148, 182)  // Button, Checked Tracking, Back 2
+
+        private static readonly Color[] _appButtonTrack =
+        {
+            Color.FromArgb(255, 251, 230),
+            Color.FromArgb(248, 230, 143),
+            Color.FromArgb(238, 213, 126),
+            Color.FromArgb(254, 247, 129),
+            Color.FromArgb(240, 201, 41)
+        };
+
+        private static readonly Color[] _appButtonPressed =
+        {
+            Color.FromArgb(235, 227, 196),
+            Color.FromArgb(228, 198, 149),
+            Color.FromArgb(166, 97, 7),
+            Color.FromArgb(242, 155, 57),
+            Color.FromArgb(236, 136, 9)
+        };
+
+        private static readonly Color[] _buttonBorderColors =
+        {
+            Color.FromArgb(212, 212, 212), // Button, Disabled, Border
+            Color.FromArgb(163, 179, 220), // Button, Tracking, Border 1
+            Color.FromArgb(128, 141, 173), // Button, Tracking, Border 2
+            Color.FromArgb(90, 99, 122), // Button, Pressed, Border 1
+            Color.FromArgb(118, 130, 160), // Button, Pressed, Border 2
+            Color.FromArgb(136, 150, 185), // Button, Checked, Border 1
+            Color.FromArgb(174, 192, 236) // Button, Checked, Border 2
+        };
+
+        private static readonly Color[] _buttonBackColors =
+        {
+            Color.FromArgb(221, 221, 221), // Button, Disabled, Back 1
+            Color.FromArgb(236, 236, 236), // Button, Disabled, Back 2
+            Color.FromArgb(54, 64, 88), // Button, Tracking, Back 1
+            Color.FromArgb(106, 123, 164), // Button, Tracking, Back 2
+            Color.FromArgb(95, 107, 137), // Button, Pressed, Back 1
+            Color.FromArgb(54, 64, 88), // Button, Pressed, Back 2
+            Color.FromArgb(54, 64, 88), // Button, Checked, Back 1
+            Color.FromArgb(122, 137, 174), // Button, Checked, Back 2
+            Color.FromArgb(119, 134, 172), // Button, Checked Tracking, Back 1
+            Color.FromArgb(135, 148, 182) // Button, Checked Tracking, Back 2
         };
 
         #endregion
@@ -739,7 +768,7 @@ namespace Krypton.Toolkit
                 case PaletteBackStyle.GridDataCellCustom2:
                 case PaletteBackStyle.GridDataCellCustom3:
                     return state == PaletteState.CheckedNormal
-                        ? _ribbonColours[(int) SchemeOfficeColors.GridDataCellSelected] : _ribbonColours[(int) SchemeOfficeColors.PanelAlternative];
+                        ? _ribbonColours[(int)SchemeOfficeColors.GridDataCellSelected] : _ribbonColours[(int)SchemeOfficeColors.PanelAlternative];
 
                 case PaletteBackStyle.GridDataCellSheet:
                     return state == PaletteState.CheckedNormal ? _buttonBackColors[6] : _ribbonColours[(int)SchemeOfficeColors.PanelAlternative];